fix: text editor for lesson
